+++ NEWS.md
@@ -2,6 +2,7 @@
 
 - Added new vignette comparing convergence issues in WTP space models with other similar packages.
 - Added `tidy()`, `glance()`, and `augment()` methods for use with the {broom} package.
+- Added `model.matrix.logitr()`, `terms.logitr()`, and `model.frame.logitr()` methods for use with {broom.helpers}.
 
 # logitr 0.7.2